# solution/api.yml
openapi: "3.0.1"
info:
  title: Promo Code API
  version: "1.0"
servers:
  - url: https://api.promocode.com/v1
paths:
...
  /user/auth/sign-up:
    post:
      tags:
        - B2C
      summary: Регистрация нового пользователя
      description: |
        Регистрирует нового пользователя и возвращает токен доступа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "200":
          description: Пользователь успешно зарегистрирован.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IlByb2QgUHJvZG92aWNoIiwiaWF0IjoxNTE2MjM5MDIyfQ.zqNHgTZOlgb0n5KDT2kPET4dOKnkrxJP5Kz6LhIGRiY"
        "400":
          $ref: "#/components/responses/Response400"
        "409":
          $ref: "#/components/responses/EmailAlreadyRegistred"
  /user/auth/sign-in:
    post:
      tags:
        - B2C
      summary: Аутентификация пользователя
      description: |
        Вход пользователя по email и паролю для получения токена доступа. Успешная аутентификация инвалидирует ранее выданные токены (запросы по ним станут невозможны).
      requestBody:
        $ref: "#/components/requestBodies/SignIn"
      responses:
        "200":
          $ref: "#/components/responses/SignIn200"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/SignIn401"
  /user/profile:
    get:
      tags:
        - B2C
      summary: Получение профиля пользователя
      description: |
        Возвращает данные профиля текущего пользователя.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Данные профиля пользователя.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/NoAuth401"

    patch:
      tags:
        - B2C
      summary: Изменение настроек пользователя
      description: |
        Обновляет настройки текущего пользователя. Если указан новый пароль, следующие попытки аутентификации должны учитывать обновленное значение.
        
        Если значение поля не было передано (или передан null), не обновляйте данное поле.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Профиль обновлен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
  /user/feed:
    get:
      tags:
        - B2C
      summary: Получение ленты промокодов
      description: |
        Возвращает ленту промокодов с поддержкой пагинации, фильтрации и сортировки. Возвращаются промокоды, которые соответствуют настройкам таргетинга.
        Промокоды отсортированы по убыванию даты создания.

      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/LimitQueryParam"
        - $ref: "#/components/parameters/OffsetQueryParam"
        - name: category
          in: query
          schema:
            type: string
            example: коты
          description: Будут возвращены промокоды с указанной категорией.
        - name: active
          in: query
          schema:
            type: boolean
            example: true
          description: Если поле указано, будут возвращены промокоды с соответствующим значением поля `active`. Если параметр отсутствует, фильтрация по статусу активности не применяется.
      responses:
        "200":
          description: Лента промокодов.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoForUser"
          headers:
            X-Total-Count:
              $ref: "#/components/headers/XTotalCount"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
  /user/promo/{id}:
    get:
      tags:
        - B2C
      summary: Просмотр промокода по id
      description: |
        Возвращает промокод с этим id
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: История промокодов.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoForUser"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoNotFound"
  /user/promo/{id}/like:
    post:
      tags:
        - B2C
      summary: Добавить лайк к промокоду
      description: |
        Добавляет лайк к указанному промокоду. Повторный лайк не изменяет состояния, возвращается успешный ответ.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Лайк успешно добавлен.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoNotFound"

    delete:
      tags:
        - B2C
      summary: Удалить лайк с промокода
      description: |
        Удаляет лайк с указанного промокода. Если лайк не стоит, возвращается успешный ответ.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Лайк успешно удален.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoNotFound"
  /user/promo/{id}/comments:
    post:
      tags:
        - B2C
      summary: Добавить комментарий к промокоду
      description: |
        Добавляет комментарий к указанному промокоду.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  $ref: "#/components/schemas/CommentText"
              required:
                - text
      responses:
        "201":
          description: Комментарий успешно добавлен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoNotFound"

    get:
      tags:
        - B2C
      summary: Получить комментарии к промокоду
      description: |
        Возвращает список комментариев к указанному промокоду. Комментарии отсортированы по убыванию даты публикации.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/LimitQueryParam"
        - $ref: "#/components/parameters/OffsetQueryParam"
        - $ref: "#/components/parameters/Id"

      responses:
        "200":
          description: Список комментариев.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
          headers:
            X-Total-Count:
              $ref: "#/components/headers/XTotalCount"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoNotFound"
  /user/promo/{id}/comments/{comment_id}:
    get:
      tags:
        - B2C
      summary: Получить комментарий к промокоду
      description: |
        Возвращает комментарий с данным ID.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CommentId"
      responses:
        "200":
          description: Комментарий.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "404":
          $ref: "#/components/responses/PromoOrCommentNotFound"

    put:
      tags:
        - B2C
      summary: Редактировать текст комментария
      description: |
        Редактировать текст комментария с данным айди.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  $ref: "#/components/schemas/CommentText"
              required:
                - text
      responses:
        "200":
          description: Комментарий.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "403":
          $ref: "#/components/responses/NoAccessToPromoUser"
        "404":
          $ref: "#/components/responses/PromoOrCommentNotFound"

    delete:
      tags:
        - B2C
      summary: Удалить комментарий
      description: |
        Удалить комментарий с данным ID.
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/CommentId"
      responses:
        "200":
          description: Комментарий успешно удален.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "403":
          $ref: "#/components/responses/NoAccessToPromoUser"
        "404":
          $ref: "#/components/responses/PromoOrCommentNotFound"
  /user/promo/{id}/activate:
    post:
      tags:
        - B2C
      summary: Активация промокода
      description: |
        Активация промокода по его ID.
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Промокод успешно активирован.
          content:
            application/json:
              schema:
                type: object
                properties:
                  promo:
                    type: string
                    example: "sale-GOOOL-50"
        "400":
          $ref: "#/components/responses/Response400"
        "401":
          $ref: "#/components/responses/NoAuth401"
        "403":
          description: Вы не можете использовать этот промокод
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Вы не можете использовать этот промокод."
                    description: Возвращается в случае, если антифрод сервис не разрешил данному пользователю получить новый промокод.
        "404":
          $ref: "#/components/responses/PromoNotFound"
...
components:
  requestBodies:
    SignIn:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SignIn"
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PromoId"
    CommentId:
      name: comment_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/CommentId"
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
        minLength: 5
        maxLength: 300
      description: Токен доступа - "Bearer {token}".
      example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IlByb2QgUHJvZG92aWNoIiwiaWF0IjoxNTE2MjM5MDIyfQ.zqNHgTZOlgb0n5KDT2kPET4dOKnkrxJP5Kz6LhIGRiY"
    LimitQueryParam:
      name: limit
      in: query
      schema:
        type: integer
        description: Максимальное количество записей
        minimum: 0
        default: 10
    OffsetQueryParam:
      name: offset
      in: query
      schema:
        type: integer
        description: Сдвинуть начало на
        minimum: 0
        default: 0
  schemas:
...
    SignIn:
      type: object
      properties:
        email:
          $ref: "#/components/schemas/Email"
        password:
          $ref: "#/components/schemas/Password"
      required:
        - email
        - password
...
    Password:
      type: string
      description: Пароль пользователя/компании. Должен содержать латинские буквы, хотя бы одну заглавную, одну строчную, одну цифру и специальные символы.
      pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
      minLength: 8
      maxLength: 60
      example: HardPa$$w0rd!iamthewinner
...
    PromoForUser:
      type: object
      description: "Поля промокода при получении пользователем"
      properties:
        promo_id:
          $ref: "#/components/schemas/PromoId"

        company_id:
          $ref: "#/components/schemas/CompanyId"

        company_name:
          $ref: "#/components/schemas/CompanyName"

        description:
          $ref: "#/components/schemas/PromoDescription"

        image_url:
          $ref: "#/components/schemas/PromoImageURL"

        active:
          $ref: "#/components/schemas/PromoIsActive"

        is_activated_by_user:
          type: boolean
          description: Активировал ли когда-то пользователь этот промокод
          example: false

        like_count:
          type: integer
          minimum: 0
          description: Количество лайков у промокода

        is_liked_by_user:
          type: boolean
          description: Поставил ли пользователь лайк этому промокоду
          example: false

        comment_count:
          type: integer
          minimum: 0
          description: Количество комментариев у промокода
          example: 5

      required:
        - promo_id
        - company_id
        - company_name
        - description
        - image_url
        - active
        - is_activated_by_user
        - like_count
        - is_liked_by_user
        - comment_count
...
    UserFirstName:
      type: string
      minLength: 1
      maxLength: 100
      example: Мария
    UserSurname:
      type: string
      minLength: 1
      maxLength: 120
      example: Федотова
...
    User:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/UserFirstName"

        surname:
          $ref: "#/components/schemas/UserSurname"

        email:
          $ref: "#/components/schemas/Email"

        avatar_url:
          $ref: "#/components/schemas/UserAvatarURL"

        other:
          $ref: "#/components/schemas/UserTargetSettings"

      required:
        - name
        - surname
        - email
        - other
    UserRegister:
      type: object
      properties:
        password:
          $ref: "#/components/schemas/Password"
      
      allOf:
        - $ref: "#/components/schemas/User"
      
      required:
        - password
    UserPatch:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/UserFirstName"

        surname:
          $ref: "#/components/schemas/UserSurname"

        avatar_url:
          $ref: "#/components/schemas/UserAvatarURL"

        password:
          $ref: "#/components/schemas/Password"
    CommentText:
      type: string
      minLength: 10
      maxLength: 1000
      description: Текст комментария
      example: Отличное предложение, все работает
    Comment:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/CommentId"

        text:
          $ref: "#/components/schemas/CommentText"

        date:
          type: string
          format: date-time
          description: Дата и время создания комментария
          example: 2025-01-02T15:04:05Z07:00
        author:
          type: object
          properties:
            name:
              $ref: "#/components/schemas/UserFirstName"
            surname:
              $ref: "#/components/schemas/UserSurname"
            avatar_url:
              $ref: "#/components/schemas/UserAvatarURL"
...
  headers:
    XTotalCount:
      schema:
        type: integer
        example: 37
      description: Суммарное число объектов. Сервер должен передавать данный заголовок в ответах на запросы с пагинацией, чтобы клиент знал, сколько объектов существует.
  responses:
    Response400:
      description: Ошибка в данных запроса. Например, несоответствие ожидаемому формату или не несоблюдение ограничений (на длину, на допустимые символы, ...).
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Ошибка в данных запроса."
    SignIn200:
      description: Успешная аутентификация.
      content:
        application/json:
          schema:
            type: object
            properties:
              token:
                type: string
                maxLength: 300
                description: Токен доступа для авторизованных запросов. После успешной аутентификации СТАРЫЕ ТОКЕНЫ ПЕРЕСТАЮТ РАБОТАТЬ.
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IlByb2QgUHJvZG92aWNoIiwiaWF0IjoxNTE2MjM5MDIyfQ.zqNHgTZOlgb0n5KDT2kPET4dOKnkrxJP5Kz6LhIGRiY"
    SignIn401:
      description: Неверный email или пароль.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Неверный email или пароль."
    NoAuth401:
      description: Пользователь не авторизован.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Пользователь не авторизован."
...
    NoAccessToPromoUser:
      description: Комментарий или промокод не принадлежит пользователю
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Комментарий не принадлежит пользователю."
    PromoNotFound:
      description: Промокод не найден.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Промокод не найден."
    EmailAlreadyRegistred:
      description: Такой email уже зарегистрирован в системе
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Такой email уже зарегистрирован."
    PromoOrCommentNotFound:
      description: Такого промокода или комментария не существует
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Такого промокода или комментария не существует."